+++ Data/HashMap/Strict/Base.hs
@@ -23,7 +23,7 @@
 -- Many operations have a average-case complexity of /O(log n)/.  The
 -- implementation uses a large base (i.e. 16) so in practice these
 -- operations are constant time.
-module Data.HashMap.Strict
+module Data.HashMap.Strict.Base
     (
       -- * Strictness properties
       -- $strictness