+++ CHANGES.md
@@ -2,6 +2,8 @@
 
  * Add `HashMap.alterF`.
 
+ * Add `HashMap.keysSet`.
+
 ## 0.2.9.0
 
  * Add `Ord/Ord1/Ord2` instances. (Thanks, Oleg Grenrus)